+++ doc/releaseNotes.md
@@ -573,3 +573,5 @@ This file describes the main feature changes for each readout.exe released versi
 
 ## next version
 - When running from the command line, the environment variable O2_RUN can be used to set the run number. It is set to 0 by default, i.e. undefined run number.
+- Updated configuration parameters:
+  - added readout.numberOfRuns: in standalone mode, number of START/STOP cycles to execute (used for testing).